Debug

+++ commerce/warehouse/src/main/resources/schema.sql
@@ -1,9 +1,9 @@
 CREATE TABLE IF NOT EXISTS product_in_warehouse (
     id VARCHAR PRIMARY KEY,
     fragile BOOLEAN,
-    width DOUBLE,
-    height DOUBLE,
-    depth DOUBLE,
-    weight DOUBLE,
+    width FLOAT,
+    height FLOAT,
+    depth FLOAT,
+    weight FLOAT,
     quantity INTEGER
 );
